# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'Partial 1 - Rugby World Cup 2023'","description":"'Partial 1 - Rugby World Cup 2023'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'Partial 1 - Rugby World Cup 2023'"}],"paths":{"/teams":{"get":{"tags":["team-controller"],"operationId":"getTeams","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TeamResponse"}}}}}}}},"/teams/{id}":{"get":{"tags":["team-controller"],"operationId":"getTeamById","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TeamResponse"}}}}}}},"/pools":{"get":{"tags":["pool-controller"],"operationId":"getPools","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PoolDto"}}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/matches":{"get":{"tags":["match-controller"],"operationId":"getMatches","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MatchResponse"}}}}}}}},"/matches/{id}":{"get":{"tags":["match-controller"],"operationId":"getMatchById","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MatchResponse"}}}}}}},"/matches/pool/{pool}":{"get":{"tags":["match-controller"],"operationId":"getMatchesByPool","parameters":[{"name":"pool","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MatchResponse"}}}}}}}}},"components":{"schemas":{"TeamResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"country":{"type":"string"},"pool":{"type":"string"},"points":{"type":"integer","format":"int32"},"tries":{"type":"integer","format":"int32"},"yellow_cards":{"type":"integer","format":"int32"},"red_cards":{"type":"integer","format":"int32"},"world_ranking":{"type":"string"}}},"PoolDto":{"type":"object","properties":{"name":{"type":"string"},"teams":{"type":"array","items":{"$ref":"#/components/schemas/TeamDto"}}}},"TeamDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"country":{"type":"string"},"matchesPlayed":{"type":"integer","format":"int32"},"wins":{"type":"integer","format":"int32"},"draws":{"type":"integer","format":"int32"},"losses":{"type":"integer","format":"int32"},"pointsFor":{"type":"integer","format":"int32"},"pointsAgainst":{"type":"integer","format":"int32"},"pointsDifferential":{"type":"integer","format":"int32"},"triesMade":{"type":"integer","format":"int32"},"bonusPoints":{"type":"integer","format":"int32"},"points":{"type":"integer","format":"int32"},"totalYellowCards":{"type":"integer","format":"int32"},"totalRedCards":{"type":"integer","format":"int32"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"MatchResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date-time"},"teams":{"type":"array","items":{"$ref":"#/components/schemas/TeamResponse"}},"stadium":{"type":"string"},"pool":{"type":"string"}}}}}}