# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/common/PoolDto.java
package ar.edu.utn.frc.tup.lciii.dtos.common;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;
@Data
@AllArgsConstructor
@NoArgsConstructor
public class PoolDto {

    private String name;
    private List<TeamDto> teams; // <Team>
}